增加ftp服务，ftpUtil支持断点续传，controller后续改进

+++ pandaz-commons/src/main/java/com/github/pandaz/commons/code/BizCode.java
@@ -50,7 +50,7 @@ public enum BizCode implements ICode {
      * @param code    编码
      * @param message 信息
      */
-    BizCode(int code, String message) {
+    BizCode(long code, String message) {
         this.code = code;
         this.message = message;
     }
